bugfix Add @Data to the class parameter,change the version number of Common to 1.0.3
