+++ pom.xml
@@ -5,7 +5,7 @@
 
   <groupId>com.aliyun</groupId>
   <artifactId>ververica-common</artifactId>
-  <version>1.0.2</version>
+  <version>1.0.3</version>
   <modelVersion>4.0.0</modelVersion>
 
   <name>ververica-common</name>
