+++ src/main/java/com/ververica/common/model/customconnector/Table.java
@@ -1,5 +1,8 @@
 package com.ververica.common.model.customconnector;
 
+import lombok.Data;
+
+@Data
 public class Table {
   String catalogName;
   String databaseName;
