+++ src/main/java/com/ververica/common/params/CreateMemberParams.java
@@ -1,7 +1,5 @@
 package com.ververica.common.params;
 
 import com.ververica.common.model.member.Member;
-import lombok.Data;
 
-@Data
 public class CreateMemberParams extends Member {}
