+++ src/main/java/com/ververica/common/params/SetPreviewSessionParams.java
@@ -4,6 +4,5 @@ import lombok.Data;
 
 @Data
 public class SetPreviewSessionParams {
-
   String previewSessionClusterName;
 }
